Keep playing on lost segments in dynamic mode
In Dynamic mode it is possible to gone some segments - thay could be gone from playlist before we take them. It could be happened becouse of PAUSE for relatively long period of time or too short playlist and segment duration on slow network. DashHandler.getNextSegmentBySequenceNumber has been corrected to detect lost segments, so that in such case DashHandler.getNextSegmentRequestFromSN returns next available in playlist segment. Becouse of lost segments, there are may be a gaps in source buffer ranges. So, the BufferController corrected to remove gaps and seek current play position.